make MPI optional through julia extension
