+++ Project.toml
@@ -5,11 +5,24 @@ version = "1.0.0"
 
 [deps]
 MPI = "da04e1cc-30fd-572f-bb4f-1f8673147195"
+MPIPreferences = "3da0fdf6-3ccc-4f1b-acd9-58baa6c99267"
 PyCall = "438e738f-606a-5dbb-bf0a-cddfbfd45ab0"
 Strided = "5e0ebb24-38b0-5f93-81fe-25c709ecae67"
 
+[extensions]
+MPIDevitoExt = "MPI"
+
 [compat]
-MPI = "0.14, 0.15, 0.16, 0.17, 0.18, 0.19, 0.20"
+MPI = "0.20.23"
+MPIPreferences = "0.1.11"
 PyCall = "1"
 Strided = "1"
 julia = "1.9"
+
+[extras]
+MPI = "da04e1cc-30fd-572f-bb4f-1f8673147195"
+Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
+Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
+
+[targets]
+test = ["Random", "Test", "MPI"]
