--- test/Project.toml
@@ -1,6 +1,0 @@
-[deps]
-MPI = "da04e1cc-30fd-572f-bb4f-1f8673147195"
-PyCall = "438e738f-606a-5dbb-bf0a-cddfbfd45ab0"
-Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
-Strided = "5e0ebb24-38b0-5f93-81fe-25c709ecae67"
-Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
