+++ test/mpitests_2ranks.jl
@@ -1,6 +1,12 @@
 using Devito, MPI, Random, Strided, Test
 
-MPI.Init()
+@info MPI.MPIPreferences.abi, MPI.MPIPreferences.binary
+
+MPIDevitoExt = Base.get_extension(Devito, :MPIDevitoExt)
+
+if !MPI.Initialized()
+    MPI.Init()
+end
 configuration!("log-level", "DEBUG")
 configuration!("language", "openmp")
 configuration!("mpi", true)
@@ -9,7 +15,7 @@ configuration!("mpi", true)
     grid = Grid(shape=n, dtype=Float32)
     b = Devito.Function(name="b", grid=grid, space_order=2)
     b_data = data_with_halo(b)
-    @test isa(b_data, Devito.DevitoMPIArray{Float32,length(n)})
+    @test isa(b_data, MPIDevitoExt.DevitoMPIArray{Float32,length(n)})
     if length(n) == 2
         @test size(b_data) == (15,14)
     else
@@ -34,7 +40,7 @@ end
     grid = Grid(shape=n, dtype=Float32)
     b = Devito.Function(name="b", grid=grid, space_order=2)
     b_data = data(b)
-    @test isa(b_data, Devito.DevitoMPIArray{Float32,length(n)})
+    @test isa(b_data, MPIDevitoExt.DevitoMPIArray{Float32,length(n)})
     @test size(b_data) == n
     b_data .= 3.14f0
 
@@ -57,7 +63,7 @@ end
     grid = Grid(shape=n, dtype=Float32)
     b = Devito.Function(name="b", grid=grid, space_order=2)
     b_data = data_with_inhalo(b)
-    @test isa(b_data, Devito.DevitoMPIArray{Float32,length(n)})
+    @test isa(b_data, MPIDevitoExt.DevitoMPIArray{Float32,length(n)})
 
     _n = length(n) == 2 ? (15,18) : (16,15,18)
 
@@ -94,7 +100,7 @@ end
     grid = Grid(shape=n, dtype=Float32)
     b = Devito.Function(name="b", grid=grid, space_order=2)
     b_data = data_with_halo(b)
-    @test isa(b_data, Devito.DevitoMPIArray{Float32,length(n)})
+    @test isa(b_data, MPIDevitoExt.DevitoMPIArray{Float32,length(n)})
 
     _n = length(n) == 2 ? (15,14) : (16,15,14)
 
@@ -133,7 +139,7 @@ end
     grid = Grid(shape=n, dtype=Float32)
     b = Devito.Function(name="b", grid=grid, space_order=2)
     b_data = data(b)
-    @test isa(b_data, Devito.DevitoMPIArray{Float32,length(n)})
+    @test isa(b_data, MPIDevitoExt.DevitoMPIArray{Float32,length(n)})
     @test size(b_data) == n
     b_data_test = zeros(Float32, n)
     if MPI.Comm_rank(MPI.COMM_WORLD) == 0
@@ -262,26 +268,26 @@ end
     MPI.Barrier(MPI.COMM_WORLD)
 end
 
-@testset "DevitoMPIArray localsize, n=$n" for n in ((5,4),(6,5,4))
+@testset "DevitoMPIArray MPIDevitoExt.localsize, n=$n" for n in ((5,4),(6,5,4))
     g = Grid(shape=n)
     f = Devito.Function(name="f", grid=g)
     h = Devito.TimeFunction(name="h", grid=g, time_order=2)
     for func in (f,h)
-        @test localsize(data(func)) == length.(Devito.localindices(data(func)))
+        @test MPIDevitoExt.localsize(data(func)) == length.(Devito.localindices(data(func)))
     end
 end
 
-@testset "DevitoMPISparseArray localsize, n=$n, npoint=$npoint" for n in ((5,4),(6,5,4)), npoint in (1,5,10)
+@testset "DevitoMPISparseArray MPIDevitoExt.localsize, n=$n, npoint=$npoint" for n in ((5,4),(6,5,4)), npoint in (1,5,10)
     g = Grid(shape=n)
     sf = SparseFunction(name="sf", grid=g, npoint=npoint)
-    @test localsize(data(sf)) == length.(Devito.localindices(data(sf)))
+    @test MPIDevitoExt.localsize(data(sf)) == length.(Devito.localindices(data(sf)))
 end
 
-@testset "DevitoMPISparseTimeArray localsize, n=$n, npoint=$npoint" for n in ((5,4),(6,5,4)), npoint in (1,5,10)
+@testset "DevitoMPISparseTimeArray MPIDevitoExt.localsize, n=$n, npoint=$npoint" for n in ((5,4),(6,5,4)), npoint in (1,5,10)
     g = Grid(shape=n)
     nt = 11
     stf = SparseTimeFunction(name="stf", grid=g, nt=11, npoint=npoint)
-    @test localsize(data(stf)) == length.(Devito.localindices(data(stf)))
+    @test MPIDevitoExt.localsize(data(stf)) == length.(Devito.localindices(data(stf)))
 end
 
 @testset "DevitoMPITimeArray, copy!, data, halo, n=$n" for n in ( (11,10), (12,11,10))
@@ -691,7 +697,7 @@ end
     grid = Grid(shape=n, dtype=Float32)
     sf = SparseFunction(name="sf", npoint=npoint, grid=grid)
     sf_coords = coordinates_data(sf)
-    @test isa(sf_coords, Devito.DevitoMPIArray)
+    @test isa(sf_coords, MPIDevitoExt.DevitoMPIArray)
     @test size(sf_coords) == (length(n),npoint)
 
     x = reshape(Float32[1:length(n)*npoint;], length(n), npoint)
@@ -728,7 +734,7 @@ end
     grid = Grid(shape=n, dtype=Float32)
     stf = SparseTimeFunction(name="stf", npoint=npoint, nt=100, grid=grid)
     stf_coords = coordinates_data(stf)
-    @test isa(stf_coords, Devito.DevitoMPIArray)
+    @test isa(stf_coords, MPIDevitoExt.DevitoMPIArray)
     @test size(stf_coords) == (length(n),npoint)
 
     x = reshape(Float32[1:length(n)*npoint;], length(n), npoint)
@@ -854,7 +860,7 @@ end
     end
     MPI.Barrier(MPI.COMM_WORLD)
     _x = data(sf)
-    @test isa(data(sf), Devito.DevitoMPISparseArray)
+    @test isa(data(sf), MPIDevitoExt.DevitoMPISparseArray)
 
     copy!(_x, x)
     x .= Float32[1:npoint;]
@@ -876,7 +882,7 @@ end
     end
     MPI.Barrier(MPI.COMM_WORLD)
     _x = data(stf)
-    @test isa(data(stf), Devito.DevitoMPISparseTimeArray)
+    @test isa(data(stf), MPIDevitoExt.DevitoMPISparseTimeArray)
 
     copy!(_x, x)
     x .= reshape(Float32[1:prod(nt*npoint);], npoint, nt)
